Rename scenario.c|h to cpp
