+++ src/openrct2-ui/input/MouseInput.cpp
@@ -33,7 +33,7 @@
 #include <openrct2/localisation/localisation.h>
 #include <openrct2/platform/platform.h>
 #include <openrct2/ride/ride_data.h>
-#include <openrct2/scenario/scenario.h>
+#include <openrct2/scenario/Scenario.h>
 #include <openrct2/world/Banner.h>
 #include <openrct2/world/map.h>
 #include <openrct2/world/scenery.h>
