+++ src/openrct2/interface/window.h
@@ -24,7 +24,7 @@
 #include "../ride/Ride.h"
 #include "../ride/TrackDesign.h"
 #include "../ride/Vehicle.h"
-#include "../scenario/scenario.h"
+#include "../scenario/Scenario.h"
 #include "../world/Park.h"
 #include "colour.h"
