+++ src/openrct2/network/Network.cpp
@@ -60,7 +60,7 @@ static sint32 _pickup_peep_old_x = LOCATION_NULL;
 #include "../interface/window.h"
 #include "../localisation/date.h"
 #include "../localisation/localisation.h"
-#include "../scenario/scenario.h"
+#include "../scenario/Scenario.h"
 #include "../util/Util.h"
 #include "../Cheats.h"
