+++ src/openrct2/platform/Crash.cpp
@@ -31,7 +31,7 @@
 extern "C"
 {
     #include "../localisation/language.h"
-    #include "../scenario/scenario.h"
+    #include "../scenario/Scenario.h"
     #include "platform.h"
 }
