+++ src/openrct2/scenario/Scenario.h
