+++ src/openrct2/world/Climate.cpp
@@ -26,7 +26,7 @@
 #include "../Game.h"
 #include "../interface/window.h"
 #include "../localisation/date.h"
-#include "../scenario/scenario.h"
+#include "../scenario/Scenario.h"
 #include "../sprites.h"
 #include "../util/Util.h"
