+++ src/openrct2/world/Particle.cpp
@@ -16,7 +16,7 @@
 
 #include "../audio/audio.h"
 #include "../core/Util.hpp"
-#include "../scenario/scenario.h"
+#include "../scenario/Scenario.h"
 #include "../paint/sprite/Sprite.h"
 #include "sprite.h"
